updated code to include a custom Context provider, confirmed working
