+++ mission-control-client/src/components/dashboard/admin-dashboard/DashboardBanner.js
@@ -1,9 +1,8 @@
 import React, { useContext } from "react";
-// Context
-import { ProductContext } from "../../../context/ProductContext"
+import { ProductContext } from '../../../context/ProductContext'
 
 const DashboardBanner = () => {
-  const { productState } = useContext(ProductContext)
+  const {productState} = useContext(ProductContext)
   return (
     <div className="dashboard-banner-container">
       <div className="dashboard-banner-head">
